Projetos_Automacao_Backend/Sucram_Video_Scene_Finder/requirements.txt
yt-dlp
youtube-search-python
google-generativeai
python-dotenv
opencv-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
youtube-transcript-api
